+++ libfreeipmi/include/freeipmi/api/ipmi-fru-inventory-device-cmds-api.h
@@ -16,7 +16,7 @@
   Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA.
 */
 /*****************************************************************************\
- *  $Id: ipmi-fru-inventory-device-cmds-api.h,v 1.6 2009-04-17 23:50:26 chu11 Exp $
+ *  $Id: ipmi-fru-inventory-device-cmds-api.h,v 1.7 2009-04-29 17:18:22 chu11 Exp $
  *****************************************************************************
  *  Copyright (C) 2007-2009 Lawrence Livermore National Security, LLC.
  *  Copyright (C) 2007 The Regents of the University of California.
@@ -53,15 +53,15 @@ extern "C" {
 #include <freeipmi/api/ipmi-api.h>
 #include <freeipmi/fiid/fiid.h>
 
-int8_t ipmi_cmd_get_fru_inventory_area_info (ipmi_ctx_t ctx,
-                                             uint8_t fru_device_id,
-                                             fiid_obj_t obj_cmd_rs);
+int ipmi_cmd_get_fru_inventory_area_info (ipmi_ctx_t ctx,
+                                          uint8_t fru_device_id,
+                                          fiid_obj_t obj_cmd_rs);
 
-int8_t ipmi_cmd_read_fru_data (ipmi_ctx_t ctx,
-                               uint8_t fru_device_id,
-                               uint16_t fru_inventory_offset_to_read,
-                               uint8_t count_to_read,
-                               fiid_obj_t obj_cmd_rs);
+int ipmi_cmd_read_fru_data (ipmi_ctx_t ctx,
+                            uint8_t fru_device_id,
+                            uint16_t fru_inventory_offset_to_read,
+                            uint8_t count_to_read,
+                            fiid_obj_t obj_cmd_rs);
 
 #ifdef __cplusplus
 }
