+++ libfreeipmi/include/freeipmi/api/ipmi-rmcpplus-support-and-payload-cmds-api.h
@@ -27,31 +27,31 @@ extern "C" {
 #include <freeipmi/api/ipmi-api.h>
 #include <freeipmi/fiid/fiid.h>
 
-int8_t ipmi_cmd_set_user_payload_access (ipmi_ctx_t ctx,
-                                         uint8_t channel_number,
-                                         uint8_t user_id,
-                                         uint8_t operation,
-                                         uint8_t standard_payload_1,
-                                         uint8_t standard_payload_2,
-                                         uint8_t standard_payload_3,
-                                         uint8_t standard_payload_4,
-                                         uint8_t standard_payload_5,
-                                         uint8_t standard_payload_6,
-                                         uint8_t standard_payload_7,
-                                         uint8_t oem_payload_0,
-                                         uint8_t oem_payload_1,
-                                         uint8_t oem_payload_2,
-                                         uint8_t oem_payload_3,
-                                         uint8_t oem_payload_4,
-                                         uint8_t oem_payload_5,
-                                         uint8_t oem_payload_6,
-                                         uint8_t oem_payload_7,
-                                         fiid_obj_t obj_cmd_rs);
+int ipmi_cmd_set_user_payload_access (ipmi_ctx_t ctx,
+                                      uint8_t channel_number,
+                                      uint8_t user_id,
+                                      uint8_t operation,
+                                      uint8_t standard_payload_1,
+                                      uint8_t standard_payload_2,
+                                      uint8_t standard_payload_3,
+                                      uint8_t standard_payload_4,
+                                      uint8_t standard_payload_5,
+                                      uint8_t standard_payload_6,
+                                      uint8_t standard_payload_7,
+                                      uint8_t oem_payload_0,
+                                      uint8_t oem_payload_1,
+                                      uint8_t oem_payload_2,
+                                      uint8_t oem_payload_3,
+                                      uint8_t oem_payload_4,
+                                      uint8_t oem_payload_5,
+                                      uint8_t oem_payload_6,
+                                      uint8_t oem_payload_7,
+                                      fiid_obj_t obj_cmd_rs);
 
-int8_t ipmi_cmd_get_user_payload_access (ipmi_ctx_t ctx,
-                                         uint8_t channel_number,
-                                         uint8_t user_id,
-                                         fiid_obj_t obj_cmd_rs);
+int ipmi_cmd_get_user_payload_access (ipmi_ctx_t ctx,
+                                      uint8_t channel_number,
+                                      uint8_t user_id,
+                                      fiid_obj_t obj_cmd_rs);
 
 
 #ifdef __cplusplus
